+++ R/get_playlists.R
@@ -10,6 +10,8 @@
 #' @param max_results Maximum number of items that should be returned. Integer. Optional. Can be between 0 and 50. Default is 50.
 #' @param page_token specific page in the result set that should be returned, optional
 #' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
+#' @param simplify Data Type: Boolean. Default is \code{TRUE}. If \code{TRUE} and if part requested is \code{contentDetails}, 
+#' the function returns a \code{data.frame}. Else a list with all the information returned.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
 #' @return captions for the video from one of the first track
@@ -21,10 +23,11 @@
 #' 
 #' # Set API token via yt_oauth() first
 #' 
-#' get_playlists(filter=c(channel_id="UChTJTbr5kf3hYazJZO-euHg"))
+#' get_playlists(filter=c(channel_id="UCMtFAi84ehTSYSE9XoHefig"))
+#' get_playlists(filter=c(channel_id="UCMtFAi84ehTSYSE9X")) # incorrect Channel ID
 #' }
 
-get_playlists <- function (filter=NULL, part="contentDetails", max_results=50, hl = NULL, page_token=NULL, ...) {
+get_playlists <- function (filter=NULL, part="contentDetails", max_results=50, hl = NULL, page_token=NULL, simplify = TRUE, ...) {
 
 	if (max_results < 0 | max_results > 50) stop("max_results only takes a value between 0 and 50")
 
@@ -38,8 +41,20 @@ get_playlists <- function (filter=NULL, part="contentDetails", max_results=50, h
 	querylist <- list(part=part, maxResults = max_results, pageToken = page_token, hl = hl)
 	querylist <- c(querylist, filter)
 
-	res <- tuber_GET("playlists", querylist, ...)
- 
- 	res
+	raw_res <- tuber_GET("playlists", querylist, ...)
+ 	
+ 	if (length(raw_res$items) ==0) { 
+    	cat("No comment information available. Likely cause: Incorrect ID. \n")
+    	if (simplify == TRUE) return(data.frame())
+    	return(list())
+    }
+
+    if (simplify==TRUE & part=="contentDetails") {
+		simple_res  <- lapply(raw_res$items, function(x) unlist(x))
+		simpler_res <- as.data.frame(do.call(rbind, simple_res))
+		return(simpler_res)
+	} 
+
+ 	raw_res
 
 }