+++ R/get_related_videos.R
@@ -24,7 +24,7 @@
 
 get_related_videos <- function (video_id=NULL, max_results=50, safe_search='none', ...) {
 
-	if (is.null(video_id)) stop("Must specify a video ID")
+	if (!is.character(video_id)) stop("Must specify a video ID")
 	if (max_results < 0 | max_results > 50) stop("max_results only takes a value between 0 and 50")
 	
 	querylist <- list(part="snippet", relatedToVideoId = video_id, type="video", maxResults=max_results, safeSearch = safe_search)
@@ -43,5 +43,5 @@ get_related_videos <- function (video_id=NULL, max_results=50, safe_search='none
 	# Cat total results
 	cat("Total Results", res$pageInfo$totalResults, "\n")
 
-	return(invisible(resdf))
+	resdf
 }