+++ R/list_channel_videos.R
@@ -4,7 +4,7 @@
 #' @param max_results Maximum number of items that should be returned. Integer. Optional. Can be between 0 and 50. Default is 50.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
-#' @return character vector
+#' @return character vector of video IDs
 #' @export
 #' @references \url{https://developers.google.com/youtube/v3/docs/channels/list}
 #' 
@@ -21,7 +21,7 @@
 list_channel_videos <- function (channel_id=NULL, max_results = 50, ...) 
 {
 
-	if (is.null(channel_id)) stop("Must specify a channel ID")
+	if (!is.character(channel_id)) stop("Must specify a channel ID")
 
     querylist <- list(id = channel_id, part="contentDetails", maxResults= max_results)
 
@@ -36,6 +36,6 @@ list_channel_videos <- function (channel_id=NULL, max_results = 50, ...)
 	# Video ids
 	vid_ids <- as.vector(unlist(sapply(vids$items, "[", "contentDetails")))
 
-    return(invisible(vid_ids))
+    vid_ids
 }