+++ R/yt_topic_search.R
@@ -18,7 +18,7 @@ yt_topic_search <- function (topic=NULL, ...) {
 
 	.Deprecated("Freebase no longer supported by Google.")
 
-	if (is.null(topic)) stop("Must specify a topic")
+	if (!is.character(topic)) stop("Must specify a topic")
 
 	yt_check_token()